refactored the components with reusable smaller components
