+++ src/components/input/index.js
@@ -1,0 +1,39 @@
+import glamorous from "glamorous";
+
+import * as COLOUR from "govuk-colours";
+import { FONT_SIZE, MEDIA_QUERIES, NTA_LIGHT } from "../../constants/index";
+
+const Input = glamorous.input(
+  {
+    boxSizing: "border-box",
+    fontFamily: NTA_LIGHT,
+    WebkitFontSmoothing: "antialiased",
+    MozOsxFontSmoothing: "grayscale",
+    fontWeight: 400,
+    textTransform: "none",
+    fontSize: FONT_SIZE.SIZE_16,
+    lineHeight: "20px",
+    width: "100%",
+    padding: "5px 4px 4px",
+    border: `2px solid ${COLOUR.BLACK}`,
+    [MEDIA_QUERIES.LARGESCREEN]: {
+      width: "50%",
+      fontSize: FONT_SIZE.SIZE_19,
+      lineHeight: "25px"
+    },
+    "[disabled]": {
+      cursor: "auto"
+    },
+    ":focus": {
+      outline: `3px solid ${COLOUR.YELLOW}`,
+      outlineOffset: 0
+    }
+  },
+  ({ errorText }) => ({
+    border: errorText
+      ? `4px solid ${COLOUR.ERROR_COLOUR}`
+      : `2px solid ${COLOUR.BLACK}`
+  })
+);
+
+export default Input;
