+++ src/components/textarea/index.js
@@ -5,84 +5,16 @@ import PropTypes from "prop-types";
 import glamorous from "glamorous";
 
 import * as COLOUR from "govuk-colours";
-import { BREAKPOINTS, SITE_WIDTH } from "../../constants/index";
+import { BREAKPOINTS } from "../../constants/index";
+
+import Label from "../label/index";
+import LabelText from "../labelText/index";
+import ErrorText from "../errorText/index";
+import HintText from "../hintText/index";
 
 const mediaQueries = {
   largeScreen: `@media only screen and (min-width: ${BREAKPOINTS.LARGESCREEN})`
 };
-
-const Label = glamorous.label(
-  {
-    display: "flex",
-    flexDirection: "column",
-    [mediaQueries.largeScreen]: {
-      maxWidth: SITE_WIDTH
-    }
-  },
-  ({ errorText }) => ({
-    borderLeft: errorText ? `4px solid ${COLOUR.ERROR_COLOUR}` : "",
-    marginRight: errorText ? `15px` : "0",
-    paddingLeft: errorText ? `10px` : "0"
-  })
-);
-
-const LabelText = glamorous.span(
-  {
-    fontFamily: '"nta", Arial, sans-serif',
-    WebkitFontSmoothing: "antialiased",
-    MozOsxFontSmoothing: "grayscale",
-    fontWeight: 400,
-    fontSize: "16px",
-    lineHeight: "1.25",
-    color: `${COLOUR.BLACK}`,
-    paddingBottom: "2px",
-    [mediaQueries.largeScreen]: {
-      fontSize: "19px",
-      lineHeight: "1.31579"
-    }
-  },
-  ({ errorText }) => ({
-    fontWeight: errorText ? 700 : LabelText.fontWeight
-  })
-);
-
-const HintText = glamorous.span({
-  fontFamily: '"nta", Arial, sans-serif',
-  WebkitFontSmoothing: "antialiased",
-  MozOsxFontSmoothing: "grayscale",
-  fontWeight: 400,
-  textTransform: "none",
-  fontSize: "16px",
-  lineHeight: "1.25",
-  color: `${COLOUR.GREY_1}`,
-  [mediaQueries.largeScreen]: {
-    fontSize: "19px",
-    lineHeight: "1.31579"
-  }
-});
-
-const ErrorText = glamorous.span(
-  {
-    display: "inline-block",
-    fontFamily: '"nta", Arial, sans-serif',
-    WebkitFontSmoothing: "antialiased",
-    MozOsxFontSmoothing: "grayscale",
-    fontWeight: 400,
-    textTransform: "none",
-    fontSize: "16px",
-    lineHeight: "1.25",
-    color: `${COLOUR.ERROR_COLOUR}`,
-    paddingTop: "4px",
-    paddingBottom: "2px",
-    [mediaQueries.largeScreen]: {
-      fontSize: "19px",
-      lineHeight: "1.31579"
-    }
-  },
-  ({ errorText }) => ({
-    fontWeight: errorText ? 700 : HintText.fontWeight
-  })
-);
 
 const TextareaField = glamorous.textarea(
   {
@@ -116,8 +48,12 @@ const TextareaField = glamorous.textarea(
 const Textarea = ({ children, hintText, errorText }) => (
   <Label errorText={errorText}>
     <LabelText errorText={errorText}>{children}</LabelText>
-    <HintText>{hintText}</HintText>
-    <ErrorText errorText={errorText}>{errorText}</ErrorText>
+    {hintText ? <HintText>{hintText}</HintText> : <span />}
+    {errorText ? (
+      <ErrorText errorText={errorText}>{errorText}</ErrorText>
+    ) : (
+      <span />
+    )}
     <TextareaField type="text" rows="5" errorText={errorText} />
   </Label>
 );
